Fiction book added, add book button added, fiction error code restored from origional
This is a change to the HTML and Python code that per Zak's request, added a book in Fiction with all descriptive information. Also, an add book button was added to make the demo easier. In addition, the no book in Fiction to cause an error code has been reverted back to original code.